# swagger.yaml
...
swagger: "2.0"
info:
  description: "API de Cadastro de Produtos - Danilo"
  version: "1.0.0"
  title: "Cadastro de Produtos"
  contact:
    email: "fabricio@veronez.dev"
paths:
...
  /stress/{elemento}/tempostress/{tempoStress}/intervalo/{intervalo}/ciclos/{ciclos}:
    put:
      description: "Faz a aplicação consumir mais recurso"
      parameters:
      - name: "elemento"
        in: "path"
        description: "Elemento pra estressar. Pode ser memory ou cpu"
        required: true
        type: "string"
      - name: "tempoStress"
        in: "path"
        description: "Tempo em segundos pra estressar o elemento"
        required: true
        type: "string"
      - name: "intervalo"
        in: "path"
        description: "Intervalo de tempo entre um stress e outro"
        required: true
        type: "string"
      - name: "ciclos"
        in: "path"
        description: "Número de ciclos de stress"
        required: true
        type: "string"                        
      responses:
        "200":
          description: "OK"        
...
  /api/produto:
    get:
      description: "Listagem de todos os produtos\n"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Resultado da busca"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Produto"
        "400":
          description: "bad input parameter"
    post:
      description: "Cadastro de um produto"
      parameters:
      - in: "body"
        name: "body"
        description: "Produto que será cadastrado"
        required: true
        schema:
          $ref: "#/definitions/Produto"
      responses:
        "200":
          description: "OK"
definitions:
  Produto:
    type: "object"
    required:
    - "categoria"
    - "nome"
    - "preco"
    properties:
      id:
        type: "string"
        format: "uuid"
        example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
      nome:
        type: "string"
        example: "Geladeira"
      preco:
        type: "number"
        example: 500.0
      categoria:
        type: "string"
        example: "Eletrodomésticos"